refactor: folder structuring
