+++ components/card/card-informasi.js
