+++ components/common/footer.js
